Split CreativeModeTabModifier off into its own class
